Updated hardhat config to JS
Given that we are implementing unit tests in JavaScript, it makes sense
to keep hardhat config consistent and expressed in JavaScript.

I have also limited Solidity version to the most recent 0.7.x that
should be more stable than 0.8.x. One of the outcomes of our previous
security audits was that it might be better to use previous, more stable
version.