Let only asset pool to mint and burn coverage tokens
The initial assumption was that the AssetPool contract will extend
UnderwriterToken and only AssetPool will call internal mint and burn
functions. This setup is how Uniswap V2 pair works. Also, it was
expected that the AssetPool will be a clone contract created by EIP1167
factory.

Although this setup would work, keeping underwriter ERC20 token
separately and not mixing AssetPool and UnderwriterToken functions into
one contract feels more elegant. Keeping them separate lets to make
clear that deposit, withdrawn, and claim functions of AssetPool operate
on the collateral token and not on the underwriter token.

Regarding EIP1167 and how AssetPool is created, creating pool with
CREATE2 opcode may be better given that the higher creation cost may pay
back later in cheaper operations on the pool. This approach needs to be
revisited so for now, it is assumed token is created in the simplest
possible way - via constructor, the same AssetPool.

Given that AssetPool and UnderwriterToken are now two separate
contracts, only AssetPool contract should be able to call public mint
and burn functions of UnderwriterToken.