(Single) Asset Pool
Asset Pool is a component of a Coverage Pool. Each Asset Pool accepts a
single ERC20 token as a collateral, and returns an underwriter token.
For example, an asset pool might accept deposits in WETH in return for
covETH underwriter tokens. Underwriter tokens represent an ownership
share in the underlying collateral of the Asset Pool.

Underwriter tokens are minted using a simple formula that originates
from Uniswap. Unlike Uniswap, we do not burn tokens on the first deposit
as the reward rate amount grows linearly over time so the Asset Pool is
not susceptible to the same attack as UNI V2 pair.