Check allowance before calling transfer
Technically, it should not matter which action is done first as if there
is not enough allowance the transaction will revert. This is how
OpenZeppelin contract is implemented. However, given that
UnderwriterToken is based on an audited Uniswap LP contract code and
that in the Uniswap contract, allowance is checked before transfer, it
makes sense to keep the order aligned.