Set reward rate for the given Asset Pool
Added a function to RewardsPool, setRewardRate, that is creating a new
instance of RewardsPoolStaking for the given AssetPool - if
RewardPoolStaking for the given AssetPool does not yet exist - and assigns
this RewardsPoolStaking a reward rate. This function can only be called by
the governance.

Separately, we will implement a change that will take into account
assigned reward rate when minting virtual reward tokens in
RewardsPoolStaking contract.