Emit event from RewardsPool.setRewardRate
This event could be helpful for UI because it easily allows to scan all
RewardsPoolStaking contracts ever created by the governance.

I have disabled reentrancy-events check for Slither. Both initialize and
setRewardRate called on RewardsPoolStaking are calls to a trusted
contract. Rearranging the code was not easily achievable so I decided
for a next-line disable.