Do not allow Auction master contract to be initialized or destroyed
The assumption was that we would initialize Auction contract during the
deployment so that it cannot be later done by a third party. This
assumption is brittle though - it is easy to forget and humans do
forget things. Consequences could be tragic. One would be able to
initialize the master contract in a way that would let them call
harikari and destroy the master contract.

From now on, we do not rely on our memory or deployment scripts but on
th contract code. It is not possible to initialize the master contract
and there is additional fuse in harikari itself, not allowing to destroy
the master contract.

Read more about similar case here:
https://blog.trailofbits.com/2020/12/16/breaking-aave-upgradeability/