Stabilize time-dependent unit tests
So far, the time-dependent unit tests used a
combination of `evm_increaseTime` and `evm_mine`.
The `evm_increaseTime` RPC method receives a
number of seconds that will be added to the
timestamp of the next block which is eventually
mined by `evm_mine` call. However, this
behavior cause occasional tests failures as
sometimes one additional second is added, probably
due to some overhead introduced by the
`evm_increaseTime` method itself.

To fix that problem, we're switching to the
Hardhat's `evm_setNextBlockTimestamp` method
which allow setting an exact timestamp for
the next block. We're calculating it by taking
the timestamp of the latest block and adding
the `time` parameter which determines the
time travel delta. This solution looks far more
stable as the next block timestamp depends only
on the timestamp of the previously mined block.