Mint reward tokens for Asset Pools without COV token staking
The mechanism previously implemented in RewardsPool contract required
underwriters to stake their COV tokens so that reward tokens could be
minted to their account. This mechanism allowed to solve problems with
underwriters joining the pool at different time, giving each one a fair
share in the rewards accumulated by the Asset Pool. On the other hand,
this mechanism did not allow for a great capital efficiency given that
COV tokens had to be staked (locked) in the contract.

The new mechanism implemented in this change does not require
underwriters to stake their COV tokens in the Rewards Pool. Reward
tokens are minted to Asset Pool all the time, based on the reward rate
assigned to the given Asset Pool by the governance.
This mechanism allows for a better capital efficiency but introduces a
problem with fair reward split share given that underwriters will be
joining the pool at different time. This problem has been moved to the
entrance and exit market component and should be solved there.

Additionally, this change makes the reward weight 1e18 precision number
so it is now possible to assign reward rates such as 1.5.