Attempting change the tests structure because of func visibility
Previously we has a couple of functions in Auctioneer that were public
types and tests were written against them. However, when we added
RiskMangerV1 and those functions became internal, they became not
visible to tests.
This is a draft to test existing Auctioneer functionality indirectly through
external funtions in RiskMangerV1.