Integration test for liquidation happy path
Introduced integration test executing a happy path for deposit
liquidation. This test allows to check if the final state is as expected
without having to introduce stubs and expose internal functions (let's
leave this hack for unit tests).

I have also removed tests for onAuctionFullyFilled from
RiskManagerV1.test.js. Those tests are not covering the core
functionality and even if we comment out purchase signer bonds and
withdraw calls, tests will still be green. The fact mappings are cleaned
up is not important at all so better to have no unit tests and rely on
integration tests than to have a false impression of a unit test
coverage for this code.