AssetPool creates RewardsPool
There is an egg and chicken problem between AssetPool and RewardsPool.
RewardsPool needs to transfer rewards only to AssetPool and AssetPool
needs to know the address of RewardsPool so that it can pull rewards
during deposit and withdrawal. The setup in which AssetPool creates
RewardsPool seems to be the best one given that reward token needs to be
the same as the collateral token.