Changed the mechanism risk manager uses for managing auction surplus
The mechanism previously proposed has two flaws. First, it is hard to
implement and requires a lot of additional bookkeeping in the contract
given that multiple auctions can be open at the same time and all of
them can generate a surplus. Secondly, it might happen that the auction
gets opened for 0.0001 TBTC because the rest was filled by the surplus.
In this scenario, the auction value is much lower than the minimum tBTC
lot size and the auction length set by the governance could make not
sense.