Two step withdrawal period from the asset pool
Withdrawal is now a two-step process with a withdrawal delay between
initiating and completing the withdrawal. It has been implemented to
avoid underwriters escaping from the pool just before the claim from the
risk manager (e.g. smart bots frontrunning risk manager/auction calls).

It takes 14 days at minimum for the underwriter to withdraw their
collateral and rewards. During that time, underwriter is still earning
rewards and their share of the pool is still a subject of a possible
coverage claim.

After 14 days withrawal delay, the underwriter has 7 days to complete
the withdrawal. After 7 days, graceful withdrawal timeout passes and
part of underwriter's rewards and collateral is slowly getting seized by
the pool - more and more over the time. This is to avoid smart
free-riders who wants to eat a cake and have a cake - be in the pool and
earn rewards but also being able to escape at any moment.

After additional 60 days (70 days from the time withdrawal delay
passed), 99% of tokens is seized by the pool and 1% of tokens is sent to
the notifier who completed the withdrawal on behalf of the underwriter.