Removed AssetPool.withdraw function and v2/EthAssetPool
In the previous commit, a two-step withdrawal process has been
implemented. That two-step process does not work with the current
implementation of EthAssetPool given that the EthAssetPool becomes the
one initiating the withdrawal and for each next withdrawal, the previous
one gets overwritten in the AssetPool contract. Two-step withdrawal
needs much more complex accounting in EthAssetPool, probably with each
WETH deposit being a separate contract. Given that we do not need
EthAssetPool for v1 release, there is no point in spending too much time
on reworking it right now.