Disabled divide-before-multiply detector for asset pool withdrawal
Disabled this detector for the case when the graceful withdrawal period
passed. Trying to multiply before division in this case makes the code
quite clumsy and does not gain that much - the number depends on the
time anyway, and most of the underwriters should complete the withdrawal
before the graceful timeout passes.