Slither made happy about event reentrancy
We now call safeTransfer on collateral token AFTER emitting events.
rewardsPool.withdraw call has been ignored given that it is a trusted
contract.

Precisely, the token will also be a trusted contract given that we go
with KEEP-only but better use the good practice everywhere than to be
sorry one day.