Get rid of the collateralization threshold
We could not determine the collateralization
threshold of a deposit being in liquidation
state as their bonds are seized at liquidation
beginning. Because of that, we can not rely
on the collateralization threshold in the risk
manager. Instead, we can use deposit's auction
value and let the risk manager rely on the bond
auction threshold.