Allow only collateral token to call receiveApproval
Only collateral token, can call receiveApproval function.

receiveApproval is needed for approveAndCall so that underwriters can
deposit tokens in one transaction. Not every token supports
approveAndCall but KEEP token that will be used for v1 of coverage pools
does.

Without requiring that only collateral token can call receiveApproval,
we are leaving a possibility for third parties to invest on behalf of
underwriters who approved the asset pool. This is especially interesting
in scenarios when the collateral token does not receive the approval in
case the caller is the spender.