Validate deposit's bond auction value
Risk manager should not open an auction for a
deposit if deposit's bond auction offers a bond
percentage lower than a `bondAuctionThreshold`.

This is because the risk manager should open a
coverage pool auction for only those deposits
that nobody else is willing to purchase. It is
expected that arbitrage bots will purchase signer
bonds for most of the liquidated tBTC deposits
before the risk manager will be required to
open an auction.

The `bondAuctionThreshold` parameter can be
changed by the governance after preserving
the governance delay.