Added self-deposit safeguard
Several high-profile tokens do not decrease the approval on transferFrom if the
caller is the spender. For example, UNI is not even checking the allowance if
spender == src:

function transferFrom(address src, address dst, uint rawAmount) external returns (bool) {
  address spender = msg.sender;
   // (...)

  if (spender != src && spenderAllowance != uint96(-1)) {
    // update allowance
  }

  _transferTokens(src, dst, amount);
  return true;
}

If collateralToken were one of these tokens, `receiveApproval` could be used to
mint unlimited number of COV and make investments of underwriters worthless:
NOP if depositor is the AssetPool's address.

receiveApproval(
    AssetPool_address,
    UNI.balanceOf(AssetPool_address),
    UNI_address, ""
)

calls

UNI.safeTransferFrom(
    AssetPool_address,
    AssetPool_address,
    UNI.balanceOf(AssetPool_address)
)

in _deposit and spender == src and COV tokens can be minted.

This attack was fixed in c205f56 but we are adding an additional safeguard in
_deposit making sure Asset Pool can not self-deposit as there is no scenario
in the system where that would happen.