Do not use hardcoded dates in ERC20WithPermit unit tests
We used hardcoded dates at the very beginning when we were not sure how
to best get the last block's timestamp from Hardhat. provider.getBlock was
very time-consuming and tests were taking ages to execute. Now we have
lastBlockTime that returns result quickly so we can use it instead of
hardcoding dates.

As a side node, our tests started to fail recently on feature branches
because the number of increaseTime calls executed in tets moved the date
beyond Jan 1 2025 when permit unit tests were executed. This problem
is gone now since we evaluate dates for permit based on the last
block's time.