Simplified withdrawal flow
The withdrawal flow has been simplified into more user-friendly form.

It takes 14 days at minimum for the underwriter to withdraw their collateral and
rewards. After 14 days from initiating the withdrawal, the underwriter has 2
days to complete the withdrawal. After 2 days, the withdrawal timeout elapses
and the underwriter has to initiate the withdrawal again and wait for the
entire withdrawal delay one more time.

Withdrawal delay and withdrawal timeout are currently not governable but they
are going to be made governable in future, in a separate change set.

Also, their initial value of 14 and 2 days can be adjusted.