Swapped notifyLiquidation assertions
In case the deposit was liquidated from surplus, the check on whether
the risk manager was notified about the deposit will not be executed. In
the previous version, the end result would be the same but that check would
pass which is fine, because we'll fail on the very next state check.
However... this could be confusing, so swapping requires.