Coverage pool uses governance delay from the asset pool
When approving new risk manager, use the governance delay from the asset
pool. This delay is enough for all underwriters to withdraw their
positions in case they do not agree with the change. Using a hardcoded,
30 days delay was fine as long as the timeout/delay for withdrawal were
hardcoded in the asset pool. Now, given that they can be altered,
coverage pool can not use a hardcoded value.