Increased withdrawal delay to 21 days
We were looking at historical liquidations and 14 days withdrawal delay
could not be enough in all cases. Of course those are just historical
data and there is no guarantee future will look similar but if something
already happened there is a chance it will happen again and we should be
prepared.

All liquidations we had on mainnet can be reduced to three cases:
- liquidations around ~Nov 5th,
- liquidations around ~Nov 17th,
- liquidations around ~Dec 26th.

If I were coverage pool underwriter having a strategy to withdraw all my
funds when it starts getting risky setting the threshold at 135%
collateralization level, two week withdrawal period would not stop me
from escaping the pool for Nov 17th liquidation. We can not stop people
from withdrawing but the whole point of a coverage pool is to be with
the pool when it is sunny outside, everyone is happy, and earning
rewards, and be with the pool when the storm is coming so that there is
that last-chance insurance. We do not want to make it easy for
underwriters trying to escape when they see a small cloud on the sky
and being interested in a free-ride on everyone else's shoulders. Three
weeks withdrawal delay feels like a good tradeoff.