Emit CovrageClaimed event from AssetPool.claim
We were not emitting an event from AssetPool when a portion of tokens
was claimed as a coverage. Having this event in the pool can help to
calculate historical APY without having to look at risk manager
implementation. This is handy because we may have multiple
implementations of a risk manager in the future.