Added AssetPool.totalValue function
The function returns the current collateral token balance of the asset
pool plus the reward amount (in collateral token) earned by the asset
pool and not yet withdrawn to the asset pool.

This function will be helpful for dApp to compute the current TVL of the
asset pool without having to query asset pool and rewards pool
separately.