Update the `paths` hardhat config
All `keep-network` packages store the contract artifacts in the
`artifacts` dir. To be consistent we need to update the paths hardhat
config. The default directory where build artifacts are stored is
`./artifacts`.  To avoid conflicts and be able to create a symbolic link
we set the artifacts path to `./build` in the `hardhat.config.js` file.