Update the asset pool deployment script
The `RewardsPool` contract is created in the `AssetPool` constructor.
With the `06_deploy_rewards_pool.js` script this contract will be
deployed twice. So to avoid confusion about which contract to use and
double deployment of the `RewardsPool` contract we can save the
`RewardsPool` deployment artifact in the` deployments` dir via the
`save` function from the `hardhat-deploy` plugin.