Handle external keep-network contract dependencies
We are expecting keep-network contracts deployment artifacts to be published to NPM package
registry. Here we pull the dependencies from the NPM for keep-core and
tbtc packages.

Artifacts for local development network are expected to be migrated
before running coverage-pools deployment and to be stored under:
`../keep-core/solidity/build/contracts` and
`../tbtc/solidity/build/contracts`.

We're updating the deployment scripts to get the external contracts. If
the external artifacts are not provided the deployment will faill for
all networks except `hardhat` internal network. For `hardhat` we will
deploy a stub.