Renamed local network to development
Renamed the network as it may better reflect the reasoning for the
network which is expected to be run on developer's host machine.

Additionally we will want to tag `development` and `hardhat` networks
with a `local` tag to use it in the deployments scripts for filtering
and don't want to cause a confusion.