Handle external predefined contract dependencies
We expect some external dependencies to be a static contracts under a
predefined addresses for each network. The first such contracts is
`UniswapV2Router`.
We added a sample artifact for this contract and updated the deployment
script.

The artifact has to be updated with correct address for a given network
and an ABI.

The deployment script is trying to fetch the external artifact but if
it's not found it will throw an error if the network is not tagged with
`local` tag. If it is a local network (`hardhat` or `development`) a
stub contract will be deployed.