Filter networks to throw errors on resolve external addresses
We assume addresses for external contracts to be provided for
deployment.
There are two types of external contracts:
1. Deployed by us in other keep-network modules which we expect
addresses to change often (keep-core, tbtc)
2. Deployed by external parties which we will have predefined addresses
for public network

For 1. we expect to deploy stubs only for hardhat internal network and
for other networks like development and public testnets we will deploy
contracts in these modules migration.

For 2. we expect to hardcode the addresses for each public network as
we're not migrating these contracts. For hardhat internal network and
also for development network we will deploy stubs. And this is why we
tagged `hardhat` and `development` networks with `local` tag.