Add a job to deploy contracts on testnet
This commit/PR adds a deployment of coverage-pools contracts on a
testnet (name of the testnet on which deployment will take place will
depend on the value of `environment` input passed when invoking
`workflow_dispatch` event). Before deploy script will run, we'll update
dependencies to upstream packages (`keep-core` and `tbtc` contracts are
needed for the deployment of `coverage-pools` contracts.)
Once contracts get deployed, we will push them to Tenderly (only
`ropsten` env curently supported) and we will publish npm package (which
will be later pulled by the Token Dashboard - it's not the scope of this
PR).
Last step in the job is an execution of `notify-workflow-completed`
action which notifies CI about successful run of the workflow and as a
result triggers deploy of any downstream modules specified in the
`ci/onfig.config.json` file.
As part of this PR we also change the conditions of execution of
`contracts-lint` and `contracts-slither` jobs - those jobs will no
longer be executed for `workflow_dispatch` event.