Adding depositWithMin() along with deposit() function to the asset pool
We decided to have 2 seperate functions reponsible for deposits in asset
pool. deposit(amount) is used when a user doesn't care about the amount
underwriter tokens minted, cause at the end of the day it will still
represent the same amount of collateral tokens. depositWithMin(amount,
min) is used when a user expects a min amount of underwriter tokens to
be minted.