Collateral token described in docs
Described in v1 documentation what is a collateral token. Also, made it
explicit collateral token can not be an arbitrary asset but needs to be
a trusted ERC20 that passed a security checklist and does not impose
any security thread. Coverage pool code is not secure against evil tokens
with, for example, reentrant malicious code.