Specify chainId for development network
For development environment we run geth based on the script in
`local-setup`. The geth instance is started with `1101` as networkId.
We want to explicitly define the ID in network configuration so Hardhat
verifies if we are connected to the correct instance. It is helpful when
we want to resolve dependencies of `keep-core` and `tbtc` that were
previously migrated on local network, as hardhat will error if it
connects to network different than expected.

If one does not run geth instance based on the `local-setup` and the
cahinId is different than `1101`, they have to update value of this
property according to their setup.