Define external contracts dependencies for mainnet
For mainnet migrations we want to use contracts deployed at specific
fixed addresses. We put the artifacts for these contracts into the
repository.

This is just one way it can be done. We could also install packages for
the specific versions and use these dependnecies:
```
yarn add @keep-network/keep-core@1.7.0 @keep-network/tbtc@1.1.0
```

I decided to do the hardcoded artifacts so we reduce possibility to pull
a wron package version or get something mixed by the yarn.