Prepare artifacts/ dir for package publication
We want to publish a pacakge were deployment artifacts are located in
`artifacts/` directory. Hardhat by default outputs them to
`deployments/<network>`. Since we're publishing separate package for
each network we can remove one level from this directory structure.

The script can be executed with:
./scripts/prepare-artifacts.sh --network <network>

In `package.json` we defined a `prepublishOnly` entry that will execute
before package is packed and published if `npm publish` command is run.
It expects the network option to be provided like in the following
example: `npm publish --dry-run --network=development`