Removed redundant balanceOf checks
In case of `initiateWithdrawal`, `safeTransferFrom` will fail in case
the withdrawing party has not enough covKEEP on their balance. covKEEP
inherits from ERC20WithPermit and behaves correctly so there is no risk
here.

In case of `upgradeToNewAssetPool`, `burnFrom` in the underwriter token
fails in case the upgrading party has not enough on their balance.
covKEEP inherits from ERC20WithPermit and behaves correctly so there is
no risk here.

To make the revert cause in this case more clear, I have switched the
order of `burnFrom` and `depositFor` calls. In case there is not enough
tokens on underwriter's balance, it is better to revert saying burn
amount exceeds balance instead of reverting with transfer amount exceeds
balance in the new asset pool, in case the new underwriter exceeded the
amount so significantly that the transfer amount exceeds pool's balance.