Simplified offerTaken callback, removed isOpen function of Auction
isOpen does not really makes sense given that it always returns true.
When function is fully filled, it is getting destroyed so it is not
possible to call any function on it.

offerTaken was simplified so that another ping-pong call between
Auctioneer and Auction could get eliminated. Instead of calling
isOpenFunction, Auctioneer returns information if the auction was fully
filled as one of offerTaken function parameters.