Add workflow deploying contracts on hardhat network
We're adding a workflow which will publish `coverage-pools` contracts
deployed on the `hardhat` network, so they could be later used when
building Token Dashboard on Testnet. The artifacts published by the job
will be used during verification of changes to the Dashboard (when
`Token Dashboard / Testnet` gets triggered by `push` or `pull_request`
event). We're deploying contracts on the `hardhat` network and saving
the artifacts (`--write true`) in the `deployments/hardhat`. The
`prepare-artifacts` script executed during the `npm publish ...` step
will copy the deployed contracts to the `/artifacts` folder.
The addresses of the contracts deployed on `hardhat` network won't
make sense on the Testnet, but that's not a problem, as they're not
used.

The `-dev` suffix will be used in the name of the published package.
The workflow will be triggered every time a change to contracts or
dependencies gets pushed to `main`.

Note: we're temporarily publishing to NPM registry using `--dry-run`
flag. The flag will be removed before merging of the changes
to the `main` branch.