Verify contracts on Etherscan
We're adding steps veryfiing migrated contracts on Etherscan. Steps are
added to the deployment scripts. Verification is done using the
`@nomiclabs/hardhat-ethers` hardhat plugin. After the deployment steps
we wait with the verification for 5 confirmations of contract deployment
transaction in order to give the deployment time to propagate to the
backend.