Verify contracts on Etherscan
We're adding a step verifying the migrated contracts on Etherscan. We're
using `etherscan-verify` plugin. In order for the plugin not to pick up
wrong files for verification we have to remove contracts of the
dependent projects stored in the `node_modules`.
We're also adding a step which executes in case of failures during the
contracts verifications. The step adds a warning annotation to the
GitHub page with the workflow run details.
Some info regarding configuration of `annotations.json` and
`github-action-create-annotations` action:
1. Values of `paths`, `start_line`, `end_line`, `start_column` and
   `end_column` don't seem to have much effect on how the annotation is
   displayed for workflows triggered by the `workflow_dispatch` event
   (they matter more for workflows triggered by `pull_request`, but we
   only annotate for `workflow_dispatch`).
2. We were getting errors when `end_line` was configured to different
   value than `start_line`.
3. The value of the `check-name` parameter of the action needs to be the
   same as the name of the job we want to annotate. Otherwise the
   annotation may get assigned to the random job (even in different
   workflow).