Respect semver when resolving latest contracts
In `Resolve latest contracts` step we were using `yarn add` command to
download the packages with the dependencies. But this command installs
the latest package (if exact package is not provided), ignoring the
semver config provided in the `package.json` file. Instead, we should
use `yarn upgrade`, which updates the dependencies to their latest
version based on the semver range specified in the `package.json`.
We can use the `yarn add` also if the command is followed by the package
name & version which does not fit the semver range - in that case semver
gets ignored and the indicated package gets installed.