Change withdrawalTimeout from 15 to 20 minutes
We've been using modified values of `withdrawalTimeout` and
`withdrawalDelay` parameters to produce an npm package that is going to
be used for testing (original values from `main` are too long for
testing purposes). We're now adjusting the `withdrawalTimeout` again to
even better suit the needs of the testers.
Note: changes are not intended for merge to `main`!