Fix deployment of UniswapV2Router on `development` network
Deployment of UniswapV2Router contract on `deployment` network was
failing with `deployed UniswapV2Router contract not found` error. This
was due to incorrect condition for throwing that error. After the fix
the error will be thrown if contract is missing and network is not
tagged as `local` (or if contract is missing and network has forking
enabled). In case of missing contract on `local`-tagged
network, stub contract will be deployed.