Do not allow the total value of collateral locked to exceed 2^96 - 1
In the previous commits we limited the amount of collateral locked to 2^96 - 1
and added support for governance checkpoints to UnderwriterToken contract.
Unfortunately, the validation in deposit function was not entirely correct.
Instead of validating that a single deposit can not exceed 2^96 - 1, we should
expect that the total value locked in the pool, including all previous deposits
plus the amount that is about to be deposited does not exceed 2^96 - 1. This way
we better guard that no more than 2^96 - 1 underwriter token is minted. With the
previous version it was possible to perform multiple deposits and bypass the
validation check. This would still fail later on UnderWriterToken's
beforeTokenTransfer check but that validation was not entirely correct.