Simplified max supply COV token validation
We do not need to validate the amount of minted tokens in AssetPool
contract because it is already validated in
`UnderwriterToken.beforeTokenTransfer`.

What's more, `beforeTokenTransfer` logic was simplifed by removing
unneccesary `SafeCast` call. If we make sure the `maxSupply()` is
not exceeded when minting, then there is no way it can be exceeded
when burning or transferring.