DAO support for coverage pools
`CoveragePool` exposes now `getPastVotes` function that can be used
by DAO components to calculate the voting power of coverage pool
underwriters.

The voting power is proportional to the amount of collateral tokens
the asset pool holds and underwriter's share of the pool.

Underwriter needs to first delegate their voting power in COV.