Use `development` instead of `dev` to tag published npm package
We're changing the name to unify tags used during publish step in
various CI workflows. In most places we use `development` tag for
tagging packages meant for development environment, so we're
changing the tagging here to follow the same pattern.
In previous change to this line I misidentified prevalent tag and
wrongly changed it from `development` to `dev`.