Remove `@threshold-network/solidity-contracts` before etherscan verification
In order for the `etherscan-verify` plugin not to pick up wrong files
for verification we have to remove contracts of the dependent projects
stored in the `node_modules`. We've been doing that for
`@keep-network/keep-core` and `@keep-network/tbtc` but some time ago a
new dependency was added, to `@threshold-network/solidity-contracts` and
we need to remove its files as well.