Fix @tenderly/hardhat-tenderly version to `1.0.12`
In previous commit we've configured the SemVer range for the module to
`>=1.0.12 <1.2.0`, which bumped up the `@tenderly/hardhat-tenderly` version in
`yarn.lock` to `1.1.6`, also bumping up some sub-dependencies. This proved to be
problematic, as it caused various failures of some unit tests. Probably the
error was related to the bumping of `hardhat-deploy` to `0.11.22`.
As we don't have time to investigate the issues deeper, we're going back to
using `1.0.12` version, which has been proven to work ok in the past.