Allow the risk manager to seize the exact amount
According to the recent DAO proposal (TIP-043), the Treasury Guild multisig
should be able to claim coverage from the pool. In v1, the coverage was claimed
by an automated `RiskManagerV1` running an auction and seizing (0,1] portion of
the pool. This approach was fine for an automated dutch auction but for v2, the
Risk Manager (Treasury Guild Multisig) should be able to claim an exact amount.

For this reason, the existing `seizeFunds` has been renamed to `seizePortion`
and a new function, `seizeAmount` is introduced. `seizeAmount` allows the
authorized Risk Manager to claim an exact amount of coverage.

See https://forum.threshold.network/t/tip-043-coverage-pool-migration/465