Batched Phased Escrow
BatchedPhasedEscrow was used by Keep Operations Team and is still used by
Threshold Treasury Guild to distribute rewards to v1 KEEP Coverage Pool. It is
a neat mechanism allowing to send tokens to the escrow every few months, and
then allocate rewards every week, even from a separate address.

Since this mechanism is something the DAO is familiar with, we want to use it
for T Coverage Pool as well.

BatchedPhasedEscrow code lies in the archived solidity-v1 directory of
keep-network/keep-core.
Since we want to deploy a new instance of the BatchedPhasedEscrow for T token
with T Coverage Pool as a beneficiary, the contract code has been copied
UNCHANGED to this repository. Unit tests also have been copied as-is, adjusting
them to ethers framework used in this repository.