Adjust workflow deploying contracts to support Goerli testnet
Previously we've been deploying coverage-pools on Ropsten. As this testnet
became deprecated, we're switching to the Goerli testnet. The hardhat config for
that network has already been added in a separate commit/PR. Now we need to
adjust CI jobs to be in line with that config. In this commit we modify the way
we get the contract owner account's private key - we use an environment variable
that can store multiple accounts and we fill its value with the value of a
`GOERLI_ETH_CONTRACT_OWNER_PRIVATE_KEY` GH secret.
We also update value of the `CHAIN_API_URL` variable, as we've changed the name
of the secret that holds the information about the chain API URL.