Incorporate workflow in the v2 of the CI flow
The `Solidity` workflow for `coverage-pools` was part of the no longer used v1
of the CI flow. As in the v2 of the system we changed the structure of the
dependencies between modules and added soem new modules, we need to adjust the
position of the `coverage-pools` module in the v2 of the CI flow. In a sepaarate
repository (`@keep-network/ci`) we're configuring the
`github.com/keep-network/coverage-pools` module to execute after
`github.com/threshold-network/solidity-contracts` module. In this PR we switch
the workflow to use the `v2` of the CI flow and change the way how we resolve
latest contracts. When workflow will be executed with the `goerli` `environment`
input (which is the assumed behavior), we'll use the latest `goerli`-tagged
`keep-core` and `tbtc` packages. We'll determine the version of the
`@threshold-network/solidity-contracts` package based on the value passed in the
`upstream_builds` input.
As part of this PR we've also adjusted the Hardhat config to use the artifacts
of the `@threshold-network/solidity-contracts` module.

Tmp