Verify contracts in Etherscan using `@nomiclabs/hardhat-etherscan`
After deploying the `coverage-pools` contracts we want to verify them in the
Etherscan, so that their source code could be easily browsed there.
Historically, in our other modules, we've been using either the
`@nomiclabs/hardhat-etherscan` or `hardhat-deploy` plugin to run the
verification.
Here in the `coverage-pools` module, the config we used was inconsistent and it
contained elements of both plugins' configs.
In this commit we sort the config out - from now on we'll use only the
`@nomiclabs/hardhat-etherscan` plugin to verify `coverage-pools` contracts on
Etherscan. We already have the plugin defined as the project dependency and we
have the deployments scripts configured to perform the verification. We're
adding passing of the `ETHERSCAN_API_KEY` secret as an environment variable
in the deployment step in the CI and removing an obsolete configuration related
to the `hardhat-deploy` plugin.