Update dependencies to `solidity-contracts`
We have published new `@threshold-network/solidity-contract` package. This
package no longer includes the `prepare-dependencies.sh` script which was
causing random failures during `yarn install`. As in some CI jobs we install the
dependencies based on the lockfile (with the `--frozen-lockfile` flag), we need
to update the dependencies in the lockfile so that the new, improved packages
would be used.