Support deployment on Sepolia
The Görli testnet currently used by Threshold/Keep for development purposes is
planned to become deprecated with the end of year 2023. The testnet that was
created to replace it is called
[Holešky](https://github.com/eth-clients/holesky), however it will take some
time until it gets integrated with by some of the projects we rely on.
As a solution, we decided to switch first to another testnet that is currently
live - Sepolia. This testnet's EOL is planned for 2026, which gives us plenty of
time to move to Holešky before Sepolia gets deprecated.
Until Görli is not dead we want to support both testnets, which is what we did
in some of the projects already. In `coverage-pools` however we were missing
some Görli configuration. So for now, for consistency between projects, we've
added the missing part for Görli support and also added full support of Sepolia
testnet.
Some other changes:
- we've removed the `contracts-etherscan-verification` job from CI (it was
  needed when `hardhat-deploy` was used to verify contracts, but in
  https://github.com/threshold-network/solidity-contracts/pull/125 we switched
  to using `@nomiclabs/hardhat-etherscan` plugin
- we've updated the versions of some of the actions used in our CI, as the
  previously set versions were running on `node14` which will be no longer
  supported by GH at some point in the future